Finished adjustments from meeting on 18 Nov 2025
Minutes in overleaf file. To summarize: 1) Rearranged order of functions to "load data>background>intercalib>radialscale". Changed the resampling and background calculations for uncertainties to use a root mean square. Removed the standard deviation in the background calculation. Moved all current excess functions out of the notebook. Next is the gaussian fits.