Updated tool to use a class for the event instead
All the functions and properties are confined inside the 'Event' class. Gaussian fitting is completed, last steps to complete are the final results plot, the gif of the gaussian fits, and other minor changes (such as adapting to different coordinate systems and shifting the longitudinal axis to loop around in cases of events on the dark side). The last few functions also need to be moved to the external py file.